+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is used to list changes made in each version of the cinc-omnibus cookb
 
 ## Unreleased
 
+- Remove pkg-config-lite from omnibus-toolchain binary path
+- Install pkgconf from distro as a replacement
+
 ## 1.2.1 - *2024-12-26*
 
 - Use cinc omnibus-toolchain for RHEL-10